Remove android:supportsRtl from the manifest file
Declares whether your application is willing to support right-to-left
(RTL) layouts. If set to true and targetSdkVersion is set to 17 or
higher, various RTL APIs will be activated and used by the system so
your app can display RTL layouts. If set to false or if targetSdkVersion
is set to 16 or lower, the RTL APIs will be ignored or will have no
effect and your app will behave the same regardless of the layout
direction associated to the user's Locale choice (your layouts will
always be left-to-right).

The default value of this attribute is `false`.

This attribute was added in API level 17.

Ref: http://developer.android.com/guide/topics/manifest/application-element.html